Add files via upload
This markdown contains the main functions and surrounding code for merging the contacts extracted from the Lamins Hi-C replicates. There are 6 replicates for each condition and 3 conditions. The Hi-C data was given to us as .hic files, not FASTQs. We were not able to merge the replicates in earlier steps in the analysis as a result. This code is for use on a HPC cluster or cloud.